ARTEMIS-127 Fix Array.toString(), nonatomic update on volatiles

+++ tests/activemq5-unit-tests/src/test/java/org/apache/activemq/store/kahadb/KahaDBFastEnqueueTest.java
@@ -238,8 +238,8 @@ public class KahaDBFastEnqueueTest {
     public void testRollover() throws Exception {
         byte flip = 0x1;
         for (long i=0; i<Short.MAX_VALUE; i++) {
-            assertEquals("0 @:" + i, 0, flip ^= 1);
-            assertEquals("1 @:" + i, 1, flip ^= 1);
+            assertEquals("0 @:" + i, 0, flip ^= (byte) 1);
+            assertEquals("1 @:" + i, 1, flip ^= (byte) 1);
         }
     }
 }
